Rename goGetImports to goGetImport
The plugin only fetches one import so having the file name plural is
misleading to anybody reading the code.

Microsoft/vscode-go#1189
